djangoTest/requirements.txt
django==2.*
numpy
scipy
pandas
seaborn
scikit-learn
scikit-surprise
nltk
twitter==1.18.*
